Safari-611.1.10/LayoutTests/fast/events/ios/command+shift+v-should-not-insert-v-expected.txt

Tests that Command + Shift + v does not insert v and that pressing a does insert an a.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.event.data is "a"
PASS successfullyParsed is true

TEST COMPLETE
